Fix tests broken by circular import fixes
